+++ Binance-Spot-API/Model/Market/TradePrice.cs
@@ -5,7 +5,7 @@ namespace Binance_Spot_API.Model.Market
 {
     public class TradePrice
     {
-        public double Price { get; set; }
-        public double Quantity { get; set; }
+        public decimal Price { get; set; }
+        public decimal Quantity { get; set; }
     }
 }
